requirements.txt
Flask==3.0.0
Pillow==10.0.0
Werkzeug==2.3.7
gunicorn==20.1.0
